fix: use intptr_t for buffer lengths, format diagnostics, ABI CI step
intptr_t: add to _C_TYPES_INT.  Convert all test-case buffer
length parameters from 'int' (4 bytes, overflows on >2B elements)
to 'intptr_t' (pointer-width, matches Py_ssize_t on 64-bit).
Generated code already emits Py_ssize_t expressions for buffer
lengths; the wider C parameter silences MSVC C4244 truncation
warnings and is correct for large buffers.

default_raise: include each buffer's actual format string in
the error message via PyErr_Format, so dispatch failures show
what format was received (e.g. 'got format=<I').

CI: add ABI diagnostic step before the build, printing sizeof(long),
sizeof(void*), and ctypes buffer format characters for all types.

Verified: Linux test suite 23/23 pass, 7-bit ASCII clean.

+++ tests/cases/arraysum/arraysum.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 /* arraysum.c - element-wise addition of double arrays */
-extern int array_sum(const double *a, const double *b, double *result, int n);
+extern int array_sum(const double *a, const double *b, double *result, intptr_t n);
 
-int array_sum(const double *a, const double *b, double *result, int n) {
+int array_sum(const double *a, const double *b, double *result, intptr_t n) {
     int i;
     for (i = 0; i < n; i++) {
         result[i] = a[i] + b[i];
