+++ tests/cases/dot/dot.c
@@ -1,6 +1,7 @@
+#include <stdint.h>
 /* dot.c - dot product of float or double arrays */
 
-float dot_f(const float *a, const float *b, int n) {
+float dot_f(const float *a, const float *b, intptr_t n) {
     float sum = 0.0f;
     int i;
     for (i = 0; i < n; i++) {
@@ -9,7 +10,7 @@ float dot_f(const float *a, const float *b, int n) {
     return sum;
 }
 
-double dot_d(const double *a, const double *b, int n) {
+double dot_d(const double *a, const double *b, intptr_t n) {
     double sum = 0.0;
     int i;
     for (i = 0; i < n; i++) {
