+++ tests/cases/freethreading/freethread.c
@@ -1,5 +1,6 @@
+#include <stdint.h>
 /* Simple C function for free-threading test */
-int double_it(const double *in, double *out, int n) {
+int double_it(const double *in, double *out, intptr_t n) {
     for (int i = 0; i < n; i++) out[i] = in[i] * 2.0;
     return n;
 }
